Fix forksoverknives.com
Use an array of selectors to see if a page has a hit. Every selector in
the array must match. If the array matches then every element will be added
to the popup in the order it is found. There isn't much of a use case for this
outside of forksoverknives.com but their probably will be in the future.